backend/requirements.txt
fastapi
uvicorn
pytest
requests
pillow